+++ 0.56.b/graysvr/CCharSpell.cpp
@@ -2259,7 +2259,7 @@ bool CChar::Spell_CastDone()
 			};
 
 			int iGet = 0;
-			for ( int i=0; i<COUNTOF(sm_Item_Bone); i++ )
+			for ( size_t i = 0; i < COUNTOF(sm_Item_Bone); i++ )
 			{
 				if ( ! Calc_GetRandVal( 2+iGet ))
 					break;
